requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
python-pushover==0.3
requests==2.18.4
schedule==0.5.0
selenium==3.9.0
urllib3==1.22
